+++ package.json
@@ -19,6 +19,7 @@
     "axios": "^1.6.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.23.1",
     "react-virtuoso": "^4.7.9",
     "styled-components": "^6.1.8"
   },
