--- src/components/areaBar.tsx
@@ -1,14 +1,0 @@
-import * as React from "react";
-import { BarChart } from "@mui/x-charts/BarChart";
-
-export default function BasicBars() {
-  return (
-    <BarChart
-      xAxis={[{ scaleType: "band", data: ["group A", "group B", "group C"] }]}
-      series={[{ data: [4, 3, 5] }, { data: [1, 6, 3] }, { data: [2, 5, 6] }]}
-      height={300}
-      colors={["#F18419", "#1942ac", "#FF0000"]}
-      sx={{ bgcolor: "background.paper" }}
-    />
-  );
-}
