+++ src/components/views/progressionGraph.tsx
